requirements.txt
langchain == 0.2.3
langchain-community == 0.2.4
chromadb == 0.4.18
tiktoken == 0.5.2
tqdm == 4.66.1
sentence-transformers == 3.0.1
python-dotenv == 1.0.0
openai == 1.33.0
pypdf==4.2.0
unstructured[all-docs]
pillow
pydantic
lxml
matplotlib
lida == 0.0.14
python-pptx == 0.6.23
PyMuPDF == 1.24.5

